move audio-related classes to separate dir
