+++ src/widget/form/settings/avform.cpp
@@ -17,7 +17,7 @@
 #include "avform.h"
 #include "ui_avsettings.h"
 #include "src/misc/settings.h"
-#include "src/audio.h"
+#include "src/audio/audio.h"
 
 #if defined(__APPLE__) && defined(__MACH__)
  #include <OpenAL/al.h>
