+++ src/widget/widget.cpp
@@ -33,7 +33,7 @@
 #include "maskablepixmapwidget.h"
 #include "src/historykeeper.h"
 #include "src/autoupdate.h"
-#include "src/audio.h"
+#include "src/audio/audio.h"
 #include "src/platform/timer.h"
 #include "systemtrayicon.h"
 #include "src/nexus.h"
